Refactor: Create ChatMessageListCellFactory class
